Feat: schemas refacto
