+++ app/modules/amap/schemas_amap.py
@@ -5,7 +5,7 @@ from datetime import date, datetime
 
 from pydantic import BaseModel, ConfigDict
 
-from app.core.core_endpoints.schemas_core import CoreUserSimple
+from app.core.users.schemas_users import CoreUserSimple
 from app.modules.amap.types_amap import AmapSlotType, DeliveryStatusType
