+++ app/modules/booking/schemas_booking.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 from pydantic import BaseModel, ConfigDict
 
-from app.core.core_endpoints.schemas_core import CoreUserSimple
+from app.core.users.schemas_users import CoreUserSimple
 from app.modules.booking.types_booking import Decision
