+++ app/modules/flappybird/schemas_flappybird.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 from pydantic import BaseModel
 
-from app.core.core_endpoints.schemas_core import CoreUserSimple
+from app.core.users.schemas_users import CoreUserSimple
 
 
 class FlappyBirdScoreBase(BaseModel):
