+++ app/modules/raffle/models_raffle.py
@@ -3,8 +3,8 @@
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from app.core.core_endpoints.schemas_core import CoreUser
 from app.core.groups.models_groups import CoreGroup
+from app.core.users.models_users import CoreUser
 from app.modules.raffle.types_raffle import RaffleStatusType
 from app.types.sqlalchemy import Base
