+++ app/modules/raffle/schemas_raffle.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, ConfigDict
 
-from app.core.core_endpoints.schemas_core import CoreUserSimple
+from app.core.users.schemas_users import CoreUserSimple
 from app.modules.raffle.types_raffle import RaffleStatusType
